+++ onlineretailstore/src/main/java/com/vikrant/mediaocean/repository/BillRepository.java
@@ -3,7 +3,5 @@ package com.vikrant.mediaocean.repository;
 import com.vikrant.mediaocean.entity.Bill;
 import org.springframework.data.repository.CrudRepository;
 
-import javax.persistence.Id;
-
-public interface BillRepository extends CrudRepository<Bill, Long> {
+public interface BillRepository extends CrudRepository<Bill, Integer> {
 }
